IOS-3728 Remove unused code
